Add parse for R and I type

+++ src/instructions.rs
@@ -16,18 +16,18 @@ pub enum DecodedInstruction32 {
         rd: u8,
         funct3: u8,
         rs1: u8,
-        imm: u16,
+        imm: i16,
     },
     SType {
         opcode: u8,
-        imm: u16,
+        imm: i16,
         funct3: u8,
         rs1: u8,
         rs2: u8,
     },
     BType {
         opcode: u8,
-        imm: u16,
+        imm: i16,
         funct3: u8,
         rs1: u8,
         rs2: u8,
@@ -40,10 +40,11 @@ pub enum DecodedInstruction32 {
     JType {
         opcode: u8,
         rd: u8,
-        imm: u32
+        imm: i32
     },
 }
 
+#[derive(Debug, PartialEq)]
 #[allow(non_camel_case_types)]
 pub enum ParsedInstruction32 {
     add {
@@ -99,22 +100,22 @@ pub enum ParsedInstruction32 {
     addi {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     xori {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     ori {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     andi {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     slli {
         rd: Register,
@@ -134,52 +135,52 @@ pub enum ParsedInstruction32 {
     slti {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     sltiu {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     lb {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     lh {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     lw {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     lbu {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     lhu {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     sb {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
     },
     sh {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
     },
     sw {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
     },
     beq {
         rs1: Register,
@@ -196,25 +197,25 @@ pub enum ParsedInstruction32 {
     blt {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
         condition: Condition,
     },
     bge {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
         condition: Condition,
     },
     bltu {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
         condition: Condition,
     },
     bgeu {
         rs1: Register,
         rs2: Register,
-        imm: i32,
+        imm: i16,
         condition: Condition,
     },
     jal {
@@ -224,7 +225,7 @@ pub enum ParsedInstruction32 {
     jalr {
         rd: Register,
         rs1: Register,
-        imm: i32,
+        imm: i16,
     },
     lui {
         rd: Register,
